+++ flutter_inappwebview_android/lib/src/in_app_webview/in_app_webview.dart
@@ -431,10 +431,18 @@ class AndroidInAppWebViewWidget extends PlatformInAppWebViewWidget {
       settings.useOnDownloadStart = true;
     }
     if ((params.shouldInterceptAjaxRequest != null ||
-            params.onAjaxProgress != null ||
-            params.onAjaxReadyStateChange != null) &&
-        settings.useShouldInterceptAjaxRequest == null) {
-      settings.useShouldInterceptAjaxRequest = true;
+        params.onAjaxProgress != null ||
+        params.onAjaxReadyStateChange != null)) {
+      if (settings.useShouldInterceptAjaxRequest == null) {
+        settings.useShouldInterceptAjaxRequest = true;
+      }
+      if (params.onAjaxReadyStateChange != null &&
+          settings.useOnAjaxReadyStateChange == null) {
+        settings.useOnAjaxReadyStateChange = true;
+      }
+      if (params.onAjaxProgress != null && settings.useOnAjaxProgress == null) {
+        settings.useOnAjaxProgress = true;
+      }
     }
     if (params.shouldInterceptFetchRequest != null &&
         settings.useShouldInterceptFetchRequest == null) {