+++ flutter_inappwebview_platform_interface/lib/src/in_app_webview/in_app_webview_settings.dart
@@ -264,14 +264,45 @@ class InAppWebViewSettings_ {
   ///Due to the async nature of [PlatformWebViewCreationParams.shouldInterceptAjaxRequest] event implementation,
   ///it will intercept only async `XMLHttpRequest`s ([AjaxRequest.isAsync] with `true`).
   ///To be able to intercept sync `XMLHttpRequest`s, use [InAppWebViewSettings.interceptOnlyAsyncAjaxRequests] to `false`.
+  ///If necessary, you should implement your own logic using for example an [UserScript] overriding the
+  ///[XMLHttpRequest](https://developer.mozilla.org/en-US/docs/Web/API/XMLHttpRequest) JavaScript object.
   ///
-  ///If the [PlatformWebViewCreationParams.shouldInterceptAjaxRequest] event or
-  ///any other Ajax event is implemented and this value is `null`,
+  ///If the [PlatformWebViewCreationParams.shouldInterceptAjaxRequest] event is implemented and this value is `null`,
   ///it will be automatically inferred as `true`, otherwise, the default value is `false`.
   ///This logic will not be applied for [PlatformInAppBrowser], where you must set the value manually.
   @SupportedPlatforms(
       platforms: [AndroidPlatform(), IOSPlatform(), MacOSPlatform()])
   bool? useShouldInterceptAjaxRequest;
+
+  ///Set to `true` to be able to listen at the [PlatformWebViewCreationParams.onAjaxReadyStateChange] event.
+  ///Also, [useShouldInterceptAjaxRequest] must be set to `true` to take effect.
+  ///
+  ///Due to the async nature of [PlatformWebViewCreationParams.onAjaxReadyStateChange] event implementation,
+  ///using it could cause some issues, so, be careful when using it.
+  ///In this case, you should implement your own logic using for example an [UserScript] overriding the
+  ///[XMLHttpRequest](https://developer.mozilla.org/en-US/docs/Web/API/XMLHttpRequest) JavaScript object.
+  ///
+  ///If the [PlatformWebViewCreationParams.onAjaxReadyStateChange] event is implemented and this value is `null`,
+  ///it will be automatically inferred as `true`, otherwise, the default value is `false`.
+  ///This logic will not be applied for [PlatformInAppBrowser], where you must set the value manually.
+  @SupportedPlatforms(
+      platforms: [AndroidPlatform(), IOSPlatform(), MacOSPlatform()])
+  bool? useOnAjaxReadyStateChange;
+
+  ///Set to `true` to be able to listen at the [PlatformWebViewCreationParams.onAjaxProgress] event.
+  ///Also, [useShouldInterceptAjaxRequest] must be set to `true` to take effect.
+  ///
+  ///Due to the async nature of [PlatformWebViewCreationParams.onAjaxProgress] event implementation,
+  ///using it could cause some issues, so, be careful when using it.
+  ///In this case, you should implement your own logic using for example an [UserScript] overriding the
+  ///[XMLHttpRequest](https://developer.mozilla.org/en-US/docs/Web/API/XMLHttpRequest) JavaScript object.
+  ///
+  ///If the [PlatformWebViewCreationParams.onAjaxProgress] event is implemented and this value is `null`,
+  ///it will be automatically inferred as `true`, otherwise, the default value is `false`.
+  ///This logic will not be applied for [PlatformInAppBrowser], where you must set the value manually.
+  @SupportedPlatforms(
+      platforms: [AndroidPlatform(), IOSPlatform(), MacOSPlatform()])
+  bool? useOnAjaxProgress;
 
   ///Set to `false` to be able to listen to also sync `XMLHttpRequest`s at the
   ///[PlatformWebViewCreationParams.shouldInterceptAjaxRequest] event.
@@ -2078,6 +2109,8 @@ as it can cause framerate drops on animations in Android 9 and lower (see [Hybri
     this.contentBlockers = const [],
     this.preferredContentMode = UserPreferredContentMode_.RECOMMENDED,
     this.useShouldInterceptAjaxRequest,
+    this.useOnAjaxReadyStateChange,
+    this.useOnAjaxProgress,
     this.interceptOnlyAsyncAjaxRequests = true,
     this.useShouldInterceptFetchRequest,
     this.incognito = false,