+++ flutter_inappwebview_platform_interface/CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 1.4.0-beta.3
 
 - Added `saveState`, `restoreState` methods to `PlatformInAppWebViewController` class
+- Added `useOnAjaxReadyStateChange`, `useOnAjaxProgress` properties to `InAppWebViewSettings`
 
 ## 1.4.0-beta.2